move data load to +page.ts, remove logs
